bugfix imageupload (deleted default image when changing pic)

+++ server/handler/fileHandler.js
@@ -30,5 +30,7 @@ exports.createFile = function(file, name, callback) {
 exports.deleteFile = function(user, callback) {
     var imageUrl = user.imageUrl.split("/");
     var filename = imageUrl[imageUrl.length-1];
-    fs.unlink(config.imagePath + filename, callback);
+    if(filename != "profilepicture.png") {
+        fs.unlink(config.imagePath + filename, callback);
+    }
 };
